move utils to patches
